api: simplify the file structure
moved contents of small files to api.go & api_test.go.
